Change the password is done

+++ app/DB/product/models.py
@@ -1,81 +1,86 @@
 from .. import db
 from bson import ObjectId
+from app.utils.utils import create_product_object
+from flask import jsonify
 
 collection = db["products"]
+extra_ingredients = [
+    "Ham",
+    "Pepperoni",
+    "Chicken",
+    "Pulled Pork",
+    "Italian Sausage",
+    "Beacon",
+    "Black Olive",
+    "Purple Onion",
+    "Mushroom",
+    "Corn",
+    "Green Pepper",
+    "Pineapple",
+    "Tomato",
+    "Tomato Cherry",
+    "Extra Cheese",
+    "BBQ Shot",
+    "Shot of Pesto",
+]
 
 
 def get_products():
-    results = collection.find()
+    results = collection.find({})
     products = []
     for re in results:
-        re["_id"] = str(re["_id"])
-        products.append(re)
+        product = create_product_object(re)
+        products.append(product)
     return {"error": False, "body": products}
 
 
 def get_products_for_name(name):
-    if name != "":
-        try:
-            results = collection.find({"name": name}).next()
-            results["_id"] = str(results["_id"])
-            return {"error": False, "body": results}
-        except StopIteration:
-            return {"error": True, "body": "We didn't find any product with that name"}
-    return {"error": True, "body": "You've to provide a product name"}
+    if name == "":
+        return {"error": True, "body": "You've to provide a product name"}
+    try:
+        results = collection.find({"name": name}).next()
+        product = create_product_object(results)
+        return {"error": False, "body": product.product()}
+    except StopIteration:
+        return {"error": True, "body": "We didn't find any product with that name"}
 
 
 def get_products_by_id(id):
     if id != "":
-        try:
-            results = collection.find({"_id": ObjectId(id)}).next()
-            results["_id"] = str(results["_id"])
-            return {"error": False, "body": results}
-        except StopIteration:
-            return {"error": True, "body": "We did not find any product with that name"}
-    return {"error": True, "body": "You've to provide a product id "}
+        return {"error": True, "body": "You've to provide a product id "}
+    try:
+        results = collection.find({"_id": ObjectId(id)}).next()
+        product = create_product_object(results)
+        return {"error": False, "body": product}
+    except StopIteration:
+        return {"error": True, "body": "We did not find any product with that name"}
 
 
 def create_product(body):
-    if "name" in body:
-        if "description" in body:
-            product_inserted = {
-                "name": body["name"],
-                "description": body["description"],
-            }
-            extra_ingredients = [
-                "Ham",
-                "Pepperoni",
-                "Chicken",
-                "Pulled Pork",
-                "Italian Sausage",
-                "Beacon",
-                "Black Olive",
-                "Purple Onion",
-                "Mushroom",
-                "Corn",
-                "Green Pepper",
-                "Pineapple",
-                "Tomato",
-                "Tomato Cherry",
-                "Extra Cheese",
-                "BBQ Shot",
-                "Shot of Pesto",
-            ]
-            product_inserted["extra_ingredients"] = extra_ingredients
-            if collection.find({"name": body["name"]}).count() == 0:
-                querry = collection.insert_one(product_inserted)
-                result = get_products_by_id(querry.inserted_id)
-                return {
-                    "error": False,
-                    "inserted_count_document": querry.acknowledged,
-                    "inserted_document": result,
-                }
-            return {"error": True, "body": "The product already exists"}
+    if "name" not in body:
+        return {"error": True, "body": "You have to provide a product name"}
+    if "description" not in body:
         return {"error": True, "body": "You have to provide a product description"}
-    return {"error": True, "body": "You have to provide a product name"}
+    if collection.find({"name": body["name"]}).count() == 0:
+        return {"error": True, "body": "The product already exists"}
+
+    product_inserted = {
+        "name": body["name"],
+        "description": body["description"],
+    }
+    product_inserted["extra_ingredients"] = extra_ingredients
+    querry = collection.insert_one(product_inserted)
+    result = get_products_by_id(querry.inserted_id)
+    return {
+        "error": False,
+        "inserted_count_document": querry.acknowledged,
+        "inserted_document": result,
+    }
 
 
 def remove_product_by_id(id):
+    if id == "":
+        return {"error": True, "body": "You have to provide a id of a product."}
     document = get_products_by_id(id)
     querry = collection.delete_one({"_id": ObjectId(id)})
     return {
@@ -87,6 +92,8 @@ def remove_product_by_id(id):
 
 
 def remove_product_by_name(name):
+    if name == "":
+        return {"error": True, "body": "You have to provide a name of a product."}
     document = get_products_for_name(name)
     querry = collection.delete_one({"name": name})
     return {
@@ -100,72 +107,35 @@ def remove_product_by_name(name):
 def update_product_by_name(name, new_body):
     old_product = collection.find({"name": name})
     if old_product.count() == 0:
-        return {"error": True, "status": "Unsuccefully", "code": 404}
-    extra_ingredients = [
-        "Ham",
-        "Pepperoni",
-        "Chicken",
-        "Pulled Pork",
-        "Italian Sausage",
-        "Beacon",
-        "Black Olive",
-        "Purple Onion",
-        "Mushroom",
-        "Corn",
-        "Green Pepper",
-        "Pineapple",
-        "Tomato",
-        "Tomato Cherry",
-        "Extra Cheese",
-        "BBQ Shot",
-        "Shot of Pesto",
-    ]
+        return {"error": True, "status": "Unsuccefully"}, 404
     old_product = old_product[0]
     updated_body = {"extra_ingredients": extra_ingredients}
-    for key in new_body.keys():
-        if not key == "name":
-            updated_body["name"] = old_product["name"]
-        else:
-            updated_body["name"] = new_body["name"]
-        if not key == "description":
-            updated_body["description"] = old_product["description"]
-        else:
-            updated_body["description"] = new_body["description"]
+
+    if "name" not in old_product:
+        updated_body["name"] = old_product["name"]
+    else:
+        updated_body["name"] = new_body["name"]
+
+    if "description" not in old_product:
+        updated_body["description"] = old_product["description"]
+    else:
+        updated_body["description"] = new_body["description"]
 
     if collection.find({"name": updated_body["name"]}).count() == 0:
         collection.update_one(
             {"name": name},
             {"$set": updated_body},
         )
-        product = collection.find({"name": updated_body["name"]})[0]
-        product["_id"] = str(product["_id"])
+        result = collection.find({"name": updated_body["name"]})[0]
+        product = create_product_object(result)
         return product
-    return {}
+    return {"error": True, "message": "The name of the product already exists"}
 
 
 def update_product_by_id(id, new_body):
     old_product = collection.find({"_id": ObjectId(id)})
     if old_product.count() == 0:
         return {"error": True, "status": "Unsuccefully", "code": 404}
-    extra_ingredients = [
-        "Ham",
-        "Pepperoni",
-        "Chicken",
-        "Pulled Pork",
-        "Italian Sausage",
-        "Beacon",
-        "Black Olive",
-        "Purple Onion",
-        "Mushroom",
-        "Corn",
-        "Green Pepper",
-        "Pineapple",
-        "Tomato",
-        "Tomato Cherry",
-        "Extra Cheese",
-        "BBQ Shot",
-        "Shot of Pesto",
-    ]
     old_product = old_product[0]
     updated_body = {"extra_ingredients": extra_ingredients}
     for key in new_body.keys():
@@ -183,7 +153,7 @@ def update_product_by_id(id, new_body):
             {"_id": ObjectId(id)},
             {"$set": updated_body},
         )
-        product = collection.find({"_id": ObjectId(id)})[0]
-        product["_id"] = str(product["_id"])
+        result = collection.find({"_id": ObjectId(id)})[0]
+        product = create_product_object(result)
         return product
-    return {}
+    return {"error": True, "message": "The name of the product already exists"}
